Use <select> for dpt and region admin filter
